优化部分单元测试代码，以规避潜在的 bug
com.dwarfeng.settingrepo.impl.service.FormatterSupportMaintainServiceImplTest
com.dwarfeng.settingrepo.impl.service.IahnNodeLocaleMaintainServiceImplTest
com.dwarfeng.settingrepo.impl.service.IahnNodeMaintainServiceImplTest
com.dwarfeng.settingrepo.impl.service.IahnNodeMekMaintainServiceImplTest
com.dwarfeng.settingrepo.impl.service.IahnNodeMessageMaintainServiceImplTest
com.dwarfeng.settingrepo.impl.service.ImageListNodeItemMaintainServiceImplTest
com.dwarfeng.settingrepo.impl.service.ImageListNodeMaintainServiceImplTest
com.dwarfeng.settingrepo.impl.service.ImageNodeMaintainServiceImplTest
com.dwarfeng.settingrepo.impl.service.LongTextNodeMaintainServiceImplTest
com.dwarfeng.settingrepo.impl.service.SettingCategoryMaintainServiceImplTest
com.dwarfeng.settingrepo.impl.service.SettingNodeMaintainServiceImplTest
com.dwarfeng.settingrepo.impl.service.TextNodeMaintainServiceImplTest

+++ settingrepo-impl/src/test/java/com/dwarfeng/settingrepo/impl/service/FormatterSupportMaintainServiceImplTest.java
@@ -14,6 +14,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import static org.junit.Assert.assertEquals;
 
@@ -55,6 +56,9 @@ public class FormatterSupportMaintainServiceImplTest {
             }
         } finally {
             for (FormatterSupport formatterSupport : formatterSupports) {
+                if (Objects.isNull(formatterSupport.getKey())) {
+                    continue;
+                }
                 formatterSupportMaintainService.deleteIfExists(formatterSupport.getKey());
             }
         }
