+++ settingrepo-impl/src/test/java/com/dwarfeng/settingrepo/impl/service/IahnNodeMaintainServiceImplTest.java
@@ -14,6 +14,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import static org.junit.Assert.assertEquals;
 
@@ -50,6 +51,9 @@ public class IahnNodeMaintainServiceImplTest {
             }
         } finally {
             for (IahnNode iahnNode : iahnNodes) {
+                if (Objects.isNull(iahnNode.getKey())) {
+                    continue;
+                }
                 service.deleteIfExists(iahnNode.getKey());
             }
         }
