+++ settingrepo-impl/src/test/java/com/dwarfeng/settingrepo/impl/service/IahnNodeMekMaintainServiceImplTest.java
@@ -15,6 +15,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
+import java.util.Objects;
+
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 
@@ -59,9 +61,13 @@ public class IahnNodeMekMaintainServiceImplTest {
             testIahnNodeMek = iahnNodeMekMaintainService.get(iahnNodeMek.getKey());
             assertEquals(BeanUtils.describe(iahnNodeMek), BeanUtils.describe(testIahnNodeMek));
         } finally {
-            iahnNodeMekMaintainService.deleteIfExists(iahnNodeMek.getKey());
+            if (Objects.nonNull(iahnNodeMek.getKey())) {
+                iahnNodeMekMaintainService.deleteIfExists(iahnNodeMek.getKey());
+            }
 
-            iahnNodeMaintainService.deleteIfExists(iahnNode.getKey());
+            if (Objects.nonNull(iahnNode.getKey())) {
+                iahnNodeMaintainService.deleteIfExists(iahnNode.getKey());
+            }
         }
     }
 
@@ -90,9 +96,13 @@ public class IahnNodeMekMaintainServiceImplTest {
 
             assertFalse(iahnNodeMekMaintainService.exists(iahnNodeMek.getKey()));
         } finally {
-            iahnNodeMekMaintainService.deleteIfExists(iahnNodeMek.getKey());
+            if (Objects.nonNull(iahnNodeMek.getKey())) {
+                iahnNodeMekMaintainService.deleteIfExists(iahnNodeMek.getKey());
+            }
 
-            iahnNodeMaintainService.deleteIfExists(iahnNode.getKey());
+            if (Objects.nonNull(iahnNode.getKey())) {
+                iahnNodeMaintainService.deleteIfExists(iahnNode.getKey());
+            }
         }
     }
 }
