+++ settingrepo-impl/src/test/java/com/dwarfeng/settingrepo/impl/service/ImageNodeMaintainServiceImplTest.java
@@ -14,6 +14,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import static org.junit.Assert.assertEquals;
 
@@ -52,6 +53,9 @@ public class ImageNodeMaintainServiceImplTest {
             }
         } finally {
             for (ImageNode imageNode : imageNodes) {
+                if (Objects.isNull(imageNode.getKey())) {
+                    continue;
+                }
                 service.deleteIfExists(imageNode.getKey());
             }
         }
