+++ settingrepo-impl/src/test/java/com/dwarfeng/settingrepo/impl/service/SettingCategoryMaintainServiceImplTest.java
@@ -14,6 +14,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import static org.junit.Assert.assertEquals;
 
@@ -55,6 +56,9 @@ public class SettingCategoryMaintainServiceImplTest {
             }
         } finally {
             for (SettingCategory settingCategory : settingCategories) {
+                if (Objects.isNull(settingCategory.getKey())) {
+                    continue;
+                }
                 settingCategoryMaintainService.deleteIfExists(settingCategory.getKey());
             }
         }
