+++ settingrepo-impl/src/test/java/com/dwarfeng/settingrepo/impl/service/TextNodeMaintainServiceImplTest.java
@@ -14,6 +14,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import static org.junit.Assert.assertEquals;
 
@@ -53,6 +54,9 @@ public class TextNodeMaintainServiceImplTest {
             }
         } finally {
             for (TextNode textNode : textNodes) {
+                if (Objects.isNull(textNode.getKey())) {
+                    continue;
+                }
                 textNodeMaintainService.deleteIfExists(textNode.getKey());
             }
         }
